requirements.txt
flask>=3.0
fonttools>=4.0
Pillow>=10.0
numpy>=1.24
